change the type of membershipNumber

+++ src/main/java/com/example/demo/web/DTO/response/MemberResponseDTO.java
@@ -5,7 +5,7 @@ import com.example.demo.entity.Member;
 import java.time.LocalDate;
 
 public record MemberResponseDTO(
-        Integer membershipNumber,
+        Long membershipNumber,
         String FullName,
         String nationality,
         LocalDate accessDate,
